+++ commonsrc/sprintf.h
@@ -1,9 +1,9 @@
 #pragma once
 
 #include <stdarg.h>
-#include "types.h"
+#include "stddef.h"
 
 int sprintf(char*, const char*, ...);
-int snprintf(char*, uint n, const char*, ...);
+int snprintf(char*, size_t n, const char*, ...);
 int vsprintf(char*, const char*, va_list);
-int vsnprintf(char*, uint n, const char*, va_list);
+int vsnprintf(char*, size_t n, const char*, va_list);
