+++ include/stdio.h
@@ -1,8 +1,7 @@
 #pragma once
 
 #include <stdarg.h>
-
-#include "types.h"
+#include "stddef.h"
 
 #define EOF  (-1)
 
@@ -14,24 +13,24 @@ extern FILE *stderr;
 
 FILE* fopen(const char* fileName, const char* mode);
 int fclose(FILE* fp);
-uint fwrite(const void* buffer, uint size, uint count, FILE* fp);
-uint fread(void* buffer, uint size, uint count, FILE* fp);
+size_t fwrite(const void* buffer, size_t size, size_t count, FILE* fp);
+size_t fread(void* buffer, size_t size, size_t count, FILE* fp);
 
 int fileno(const FILE *);
 
 int getc(void);
 int getchar(void);
 int fgetc(FILE* fp);
-char* gets_s(char*, uint max);
-char* fgets_s(char*, uint max, FILE* fp);
+char* gets_s(char*, size_t max);
+char* fgets_s(char*, size_t max, FILE* fp);
 int putchar(int);
 
 int printf(const char*, ...);
 int fprintf(FILE*, const char*, ...);
 int sprintf(char*, const char*, ...);
-int snprintf(char*, uint n, const char*, ...);
+int snprintf(char*, size_t n, const char*, ...);
 int vfprintf(FILE*, const char*, va_list);
 int vsprintf(char*, const char*, va_list);
-int vsnprintf(char*, uint n, const char*, va_list);
+int vsnprintf(char*, size_t n, const char*, va_list);
 
 void perror(const char*);
