+++ ulib/umalloc.c
@@ -10,7 +10,7 @@ typedef long Align;
 union header {
   struct {
     union header *ptr;
-    uint size;
+    unsigned int size;
   } s;
   Align x;
 };
@@ -43,7 +43,7 @@ free(void *ap)
 }
 
 static Header*
-morecore(uint nu)
+morecore(size_t nu)
 {
   char *p;
   Header *hp;
@@ -60,10 +60,10 @@ morecore(uint nu)
 }
 
 void*
-malloc(uint nbytes)
+malloc(size_t nbytes)
 {
   Header *p, *prevp;
-  uint nunits;
+  size_t nunits;
 
   nunits = (nbytes + sizeof(Header) - 1)/sizeof(Header) + 1;
   if((prevp = freep) == 0){
@@ -89,7 +89,7 @@ malloc(uint nbytes)
 }
 
 void*
-calloc(uint nbytes)
+calloc(size_t nbytes)
 {
   void *adr = malloc(nbytes);
   if (adr != 0)
